feat: Make activities page main page
