+++ src/app/activities/page.tsx
@@ -1,29 +1,5 @@
-"use client";
-
-import { Suspense } from 'react';
-import ActivityList from './ActivityList';
-import { Skeleton } from '@/components/ui/skeleton';
+import { redirect } from 'next/navigation';
 
 export default function ActivitiesPage() {
-  return (
-    <div className="container mx-auto py-8">
-      <h1 className="text-3xl font-bold mb-6">Activities</h1>
-      <Suspense fallback={<ActivitySkeleton />}>
-        <ActivityList />
-      </Suspense>
-    </div>
-  );
-}
-
-function ActivitySkeleton() {
-  return (
-    <div className="space-y-4">
-      {[...Array(5)].map((_, i) => (
-        <div key={i} className="flex flex-col gap-2">
-          <Skeleton className="h-6 w-3/4" />
-          <Skeleton className="h-4 w-1/2" />
-        </div>
-      ))}
-    </div>
-  );
+  redirect('/');
 }
